ViDiT-Q/eval/video/requirements.txt
numpy 
opencv-python 
scipy 
requests
pillow 
transformers 
tqdm 
decord 
pyyaml 
ipdb 
scikit-video 
open_clip_torch 
einops